fix: Fix Site Banner Section Move - Meeds-io/MIPs#175
