+++ layout-webapp/src/main/webapp/vue-app/common-layout/components/content/common/CellsSelectionBox.vue
