+++ layout-webapp/src/main/webapp/vue-app/common-layout/components/content/container/Section.vue
@@ -56,8 +56,8 @@
       :parent-id="parentId"
       :index="index"
       :class="`${mobileInColumns && mobileSectionColumnClass || ''}`"
+      :type="isDynamicSection ? 'section-columns' : 'section-grid'"
       class="position-relative overflow-initial layout-section-content"
-      type="section"
       no-background-style
       @hovered="hoverSection = $event && !drawerOpened">
       <template v-if="$root.movingParentId === storageId && (!isDynamicSection || $root.moveType === 'resize')" #footer>
